+++ common/src/main/java/net/arsenal/item/ArsenalWeapons.java
@@ -2,7 +2,7 @@ package net.arsenal.item;
 
 import net.arsenal.ArsenalMod;
 import net.arsenal.spell.ArsenalSpells;
-import net.arsenal.spell.ArsenalWeaponSpellTags;
+import net.arsenal.spell.ArsenalSpellGroups;
 import net.minecraft.item.Items;
 import net.minecraft.recipe.Ingredient;
 import net.minecraft.util.Rarity;
@@ -38,8 +38,10 @@ public class ArsenalWeapons {
             .translatedName("Champion's Greatsword")
             .withAdditionalSpell(ArsenalSpells.radiance_melee.id().toString())
             .lootTheme(Loot.Theme.DIVINE.toString()));
-    public static final Weapon.Entry unique_claymore_sw = add(Weapons.claymoreWithSkill(ArsenalMod.NAMESPACE, "unique_claymore_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK))
+    public static final Weapon.Entry unique_claymore_sw = add(Weapons.claymore(ArsenalMod.NAMESPACE, "unique_claymore_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK))
             .translatedName("Apolyon, the Soul-Render")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.CLAYMORE_DOUBLE_AXE.id().toString())
             .withAdditionalSpell(ArsenalSpells.rampaging_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -48,43 +50,43 @@ public class ArsenalWeapons {
     public static final Weapon.Entry unique_staff_damage_1 = add(Weapons.damageStaff(ArsenalMod.NAMESPACE, "unique_staff_damage_1", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.AMETHYST_BLOCK), List.of(SpellSchools.ARCANE.id, SpellSchools.FROST.id))
             .translatedName("Nexus Key")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_ARCANE_FROST.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_ARCANE_FROST.id().toString())
             .withAdditionalSpell(ArsenalSpells.cooldown_shot_spell.id().toString())
             .lootTheme(Loot.Theme.GENERIC.toString()));
     public static final Weapon.Entry unique_staff_damage_2 = add(Weapons.damageStaff(ArsenalMod.NAMESPACE, "unique_staff_damage_2", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK), List.of(SpellSchools.ARCANE.id, SpellSchools.FIRE.id))
             .translatedName("Antonidas's Staff of Rapt Concentration")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_ARCANE_FIRE.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_ARCANE_FIRE.id().toString())
             .withAdditionalSpell(ArsenalSpells.chain_reaction_spell.id().toString())
             .lootTheme(Loot.Theme.CRYSTAL.toString()));
     public static final Weapon.Entry unique_staff_damage_3 = add(Weapons.damageStaff(ArsenalMod.NAMESPACE, "unique_staff_damage_3", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.NETHERITE_SCRAP), List.of(SpellSchools.ARCANE.id, SpellSchools.FIRE.id))
             .translatedName("Draconic Battle Staff")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_ARCANE_FIRE.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_ARCANE_FIRE.id().toString())
             .withAdditionalSpell(ArsenalSpells.flame_cloud_spell.id().toString())
             .lootTheme(Loot.Theme.FIERY.toString()));
     public static final Weapon.Entry unique_staff_damage_4 = add(Weapons.damageStaff(ArsenalMod.NAMESPACE, "unique_staff_damage_4", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.NETHERITE_SCRAP), List.of(SpellSchools.FIRE.id, SpellSchools.FROST.id))
             .translatedName("Gargoyle's Bite")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_FIRE_FROST.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_FIRE_FROST.id().toString())
             .withAdditionalSpell(ArsenalSpells.leeching_spell.id().toString())
             .lootTheme(Loot.Theme.EVIL.toString()));
     public static final Weapon.Entry unique_staff_damage_5 = add(Weapons.damageStaff(ArsenalMod.NAMESPACE, "unique_staff_damage_5", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.AMETHYST_BLOCK), List.of(SpellSchools.ARCANE.id, SpellSchools.FROST.id))
             .translatedName("Mage Lord Cane")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_ARCANE_FROST.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_ARCANE_FROST.id().toString())
             .withAdditionalSpell(ArsenalSpells.shockwave_area_spell.id().toString())
             .lootTheme(Loot.Theme.GENERIC.toString()));
     public static final Weapon.Entry unique_staff_damage_6 = add(Weapons.damageStaff(ArsenalMod.NAMESPACE, "unique_staff_damage_6", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.NETHERITE_SCRAP), List.of(SpellSchools.ARCANE.id, SpellSchools.FROST.id))
             .translatedName("Endless Winter")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_ARCANE_FROST.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_ARCANE_FROST.id().toString())
             .withAdditionalSpell(ArsenalSpells.frost_cloud_spell.id().toString())
             .lootTheme(Loot.Theme.FROSTY.toString()));
     public static final Weapon.Entry unique_staff_damage_sw = add(Weapons.damageStaff(ArsenalMod.NAMESPACE, "unique_staff_damage_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.DIAMOND), List.of(SpellSchools.ARCANE.id, SpellSchools.FIRE.id, SpellSchools.FROST.id))
             .translatedName("Grand Magister's Staff of Torrents")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_ARCANE_FIRE_FROST.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_ARCANE_FIRE_FROST.id().toString())
             .withAdditionalSpell(ArsenalSpells.rampaging_spell.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -94,7 +96,7 @@ public class ArsenalWeapons {
             .attribute(AttributeModifier.bonus(SpellSchools.ARCANE.id, TIER_5_SPELL_POWER))
             .translatedName("Crystalline Life-Staff")
             .spellContainer(SpellContainers.forMagicWeapon())
-            .withSpellChoices(ArsenalWeaponSpellTags.STAFF_ARCANE_HEALING.id().toString())
+            .withSpellChoices(ArsenalSpellGroups.STAFF_ARCANE_HEALING.id().toString())
             .withAdditionalSpell(ArsenalSpells.radiance_spell.id().toString())
             .lootTheme(Loot.Theme.CRYSTAL.toString()));
     public static final Weapon.Entry unique_staff_heal_2 = add(Weapons.healingStaff(ArsenalMod.NAMESPACE, "unique_staff_heal_2", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.IRON_BLOCK))
@@ -124,6 +126,8 @@ public class ArsenalWeapons {
             .lootTheme(Loot.Theme.DIVINE.toString()));
     public static final Weapon.Entry unique_spear_sw = add(Weapons.spearWithSkill(ArsenalMod.NAMESPACE, "unique_spear_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK))
             .translatedName("Mounting Vengeance")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.SPEAR_GLAIVE.id().toString())
             .withAdditionalSpell(ArsenalSpells.leeching_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -139,6 +143,8 @@ public class ArsenalWeapons {
             .lootTheme(Loot.Theme.EVIL.toString()));
     public static final Weapon.Entry unique_dagger_sw = add(Weapons.daggerWithSkill(ArsenalMod.NAMESPACE, "unique_dagger_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK))
             .translatedName("Crux of the Apocalypse")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.DAGGER_SICKLE.id().toString())
             .withAdditionalSpell(ArsenalSpells.sundering_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -154,6 +160,8 @@ public class ArsenalWeapons {
             .lootTheme(Loot.Theme.FIERY.toString()));
     public static final Weapon.Entry unique_sickle_sw = add(Weapons.sickleWithSkill(ArsenalMod.NAMESPACE, "unique_sickle_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK))
             .translatedName("Thalassian Sickle")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.SICKLE_AXE.id().toString())
             .withAdditionalSpell(ArsenalSpells.swirling_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -161,6 +169,8 @@ public class ArsenalWeapons {
 
     public static final Weapon.Entry unique_longsword_sw = add(Weapons.swordWithSkill(ArsenalMod.NAMESPACE, "unique_longsword_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_INGOT))
             .translatedName("Dragonscale-Encrusted Longblade")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.ONE_HANDED_SLASHER.id().toString())
             .withAdditionalSpell(ArsenalSpells.sundering_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -176,6 +186,8 @@ public class ArsenalWeapons {
             .lootTheme(Loot.Theme.EVIL.toString()));
     public static final Weapon.Entry unique_double_axe_sw = add(Weapons.doubleAxeWithSkill(ArsenalMod.NAMESPACE, "unique_double_axe_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_INGOT))
             .translatedName("Sunreaver War Axe")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.DOUBLE_AXE_HAMMER.id().toString())
             .withAdditionalSpell(ArsenalSpells.rampaging_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -191,6 +203,8 @@ public class ArsenalWeapons {
             .lootTheme(Loot.Theme.CRYSTAL.toString()));
     public static final Weapon.Entry unique_glaive_sw = add(Weapons.glaiveWithSkill(ArsenalMod.NAMESPACE, "unique_glaive_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK))
             .translatedName("Shivering Felspine")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.GLAIVE_DOUBLE_AXE.id().toString())
             .withAdditionalSpell(ArsenalSpells.swirling_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));
 
@@ -221,6 +235,8 @@ public class ArsenalWeapons {
             .lootTheme(Loot.Theme.GENERIC.toString()));
     public static final Weapon.Entry unique_mace_sw = add(Weapons.maceWithSkill(ArsenalMod.NAMESPACE, "unique_mace_sw", Equipment.Tier.TIER_5, () -> Ingredient.ofItems(Items.GOLD_BLOCK))
             .translatedName("Archon's Scepter")
+            .spellContainer(SpellContainers.forMagicWeapon())
+            .withSpellChoices(ArsenalSpellGroups.MACE_SWORD.id().toString())
             .withAdditionalSpell(ArsenalSpells.guarding_strike_melee.id().toString())
             .lootTheme(Loot.Theme.ELVEN.toString()));