+++ common/src/main/java/net/arsenal/spell/ArsenalProjectiles.java
@@ -9,7 +9,7 @@ public class ArsenalProjectiles {
     public record Entry(Identifier id) { }
     public static final ArrayList<Entry> all = new ArrayList<>();
     public static Entry entry(String name) {
-        var entry = new Entry(Identifier.of(ArsenalMod.NAMESPACE, "projectile/" + name));
+        var entry = new Entry(Identifier.of(ArsenalMod.NAMESPACE, "spell_projectile/" + name));
         all.add(entry);
         return entry;
     }